Add Mode switcher for certificate upload
