+++ api/web/src/components/ConnectionEdit.vue
@@ -78,49 +78,102 @@
                                     </div>
                                 </div>
 
-                                <div class='card-header d-flex'>
-                                    <h3 class='card-title'>
-                                        Connection Authentication
-                                    </h3>
-                                    <div
-                                        v-if='!$route.params.connectionid || regen'
-                                        class='ms-auto btn-list'
-                                    >
-                                        <IconLogin
-                                            v-tooltip='"User Login"'
-                                            :size='32'
-                                            :stroke='1'
-                                            class='cursor-pointer'
-                                            @click='modal.login = true'
-                                        />
-                                        <IconPlus
-                                            v-tooltip='"P12 Upload"'
-                                            :size='32'
-                                            :stroke='1'
-                                            class='cursor-pointer'
-                                            @click='modal.upload = true'
-                                        />
-                                    </div>
+                                <div class='card-header'>
+                                    <h3 class='card-title'>Connection Authentication</h3>
                                 </div>
                                 <div class='card-body'>
-                                    <div class='row mt-3'>
-                                        <template v-if='!$route.params.connectionid || regen'>
-                                            <div class='col-md-6'>
-                                                <TablerInput
-                                                    v-model='connection.auth.cert'
-                                                    label='Connection Cert'
-                                                    :error='errors.cert'
-                                                    :rows='6'
-                                                />
+                                    <div class='row'>
+                                        <template v-if='isReady'>
+                                            <div class='col-12 d-flex align-items-center'>
+                                                <IconCheck :size='40' class='text-green'/>
+                                                <span class='mx-3'>Certificate Uploaded</span>
                                             </div>
-                                            <div class='col-md-6'>
-                                                <TablerInput
-                                                    v-model='connection.auth.key'
-                                                    label='Connection Key'
-                                                    :error='errors.key'
-                                                    :rows='6'
-                                                />
+                                        </template>
+                                        <template v-else-if='!$route.params.connectionid || regen'>
+                                            <div class='col-12 pb-2'>
+                                                <div
+                                                    class='btn-group w-100'
+                                                    role='group'
+                                                >
+                                                    <input
+                                                        id='creation'
+                                                        type='radio'
+                                                        class='btn-check'
+                                                        name='cert-type'
+                                                        autocomplete='off'
+                                                        :checked='type === "creation"'
+                                                        @click='type = "creation"'
+                                                    >
+                                                    <label
+                                                        for='creation'
+                                                        type='button'
+                                                        class='btn'
+                                                    >Machine User Creation</label>
+
+                                                    <input
+                                                        id='login'
+                                                        type='radio'
+                                                        class='btn-check'
+                                                        name='cert-type'
+                                                        autocomplete='off'
+                                                        :checked='type === "login"'
+                                                        @click='type = "login"'
+                                                    >
+                                                    <label
+                                                        for='login'
+                                                        type='button'
+                                                        class='btn'
+                                                    >User Login</label>
+
+                                                    <input
+                                                        id='p12'
+                                                        type='radio'
+                                                        class='btn-check'
+                                                        name='cert-type'
+                                                        autocomplete='off'
+                                                        :checked='type === "p12"'
+                                                        @click='type = "p12"'
+                                                    >
+                                                    <label
+                                                        for='p12'
+                                                        type='button'
+                                                        class='btn'
+                                                    >P12 Certificate Upload</label>
+
+                                                    <input
+                                                        id='raw'
+                                                        type='radio'
+                                                        class='btn-check'
+                                                        name='cert-type'
+                                                        autocomplete='off'
+                                                        :checked='type === "raw"'
+                                                        @click='type = "raw"'
+                                                    >
+                                                    <label
+                                                        for='raw'
+                                                        type='button'
+                                                        class='btn'
+                                                    >Raw Certificate</label>
+                                                </div>
                                             </div>
+                                            <template v-if='type === "raw"'>
+                                                <CertificateRaw
+                                                    @certs='marti($event)'
+                                                    @err='err = $event'
+                                                />
+                                            </template>
+                                            <template v-else-if='type === "p12"'>
+                                                <CertificateP12
+                                                    @certs='p12upload($event)'
+                                                    @err='err = $event'
+                                                />
+                                            </template>
+                                            <template v-else-if='type === "login"'>
+                                                <CertificateLogin
+                                                    @certs='marti($event)'
+                                                    @err='err = $event'
+                                                />
+                                            </template>
                                         </template>
                                         <template v-else>
                                             <div class='border px-3 py-3'>
@@ -136,7 +189,7 @@
                                                 <div class='d-flex justify-content-center'>
                                                     <button
                                                         class='btn btn-secondary'
-                                                        @click='regen=true'
+                                                        @click='regen = true'
                                                     >
                                                         Regenerate Certificate
                                                     </button>
@@ -153,10 +206,11 @@
                                                 />
 
                                                 <div class='ms-auto'>
-                                                    <a
+                                                    <button
+                                                        :disabled='!$route.params.connectionid && !isReady'
                                                         class='cursor-pointer btn btn-primary'
                                                         @click='create'
-                                                    >Save Connection</a>
+                                                    >Save Connection</button>
                                                 </div>
                                             </div>
                                         </div>
@@ -169,20 +223,6 @@
             </div>
         </div>
 
-        <Upload
-            v-if='modal.upload'
-            @certs='p12upload($event)'
-            @close='modal.upload = false'
-            @err='err = $event'
-        />
-
-        <LoginCertModal
-            v-if='modal.login'
-            @certs='marti($event)'
-            @close='modal.login = false'
-            @err='err = $event'
-        />
-
         <PageFooter />
     </div>
 </template>
@@ -190,13 +230,15 @@
 <script>
 import { std } from '/src/std.ts';
 import PageFooter from './PageFooter.vue';
-import AgencySelect from './util/AgencySelect.vue';
-import Upload from './util/UploadP12.vue';
-import LoginCertModal from './util/LoginCertModal.vue';
+import AgencySelect from './Connection/AgencySelect.vue';
+import CertificateP12 from './Connection/CertificateP12.vue';
+import CertificateLogin from './Connection/CertificateLogin.vue';
+import CertificateRaw from './Connection/CertificateRaw.vue';
 import {
     IconPlus,
     IconLock,
     IconLogin,
+    IconCheck,
 } from '@tabler/icons-vue';
 import {
     TablerLoading,
@@ -208,14 +250,16 @@ import {
 export default {
     name: 'ConnectionNew',
     components: {
-        Upload,
         IconPlus,
         IconLock,
         IconLogin,
+        IconCheck,
         AgencySelect,
         TablerDelete,
         TablerBreadCrumb,
-        LoginCertModal,
+        CertificateP12,
+        CertificateRaw,
+        CertificateLogin,
         TablerInput,
         TablerLoading,
         PageFooter,
@@ -224,15 +268,13 @@ export default {
         return {
             loading: true,
             regen: false,
+            type: 'raw',
             modal: {
-                login: false,
                 upload: false,
             },
             errors: {
                 name: '',
                 description: '',
-                cert: '',
-                key: ''
             },
             connection: {
                 name: '',
@@ -244,8 +286,16 @@ export default {
         }
     },
     mounted: async function() {
-        if (this.$route.params.connectionid) await this.fetch();
-        else this.loading = false;
+        if (this.$route.params.connectionid) {
+            await this.fetch();
+        } else {
+            this.loading = false;
+        }
+    },
+    computed: {
+        isReady: function() {
+            return this.connection.auth.cert.trim().length && this.connection.auth.key.trim().length
+        }
     },
     methods: {
         fetch: async function() {
@@ -255,7 +305,6 @@ export default {
             this.loading = false;
         },
         marti: function(certs) {
-            console.error(certs);
             this.connection.auth.cert = certs.cert;
             this.connection.auth.key = certs.key;
         },
@@ -276,13 +325,6 @@ export default {
             for (const field of ['name', 'description' ]) {
                 if (!this.connection[field]) this.errors[field] = 'Cannot be empty';
                 else this.errors[field] = '';
-            }
-
-            if (!this.$route.params.connectionid || this.regen) {
-                for (const field of Object.keys(this.connection.auth)) {
-                    if (!this.connection.auth[field]) this.errors[field] = 'Cannot be empty';
-                    else this.errors[field] = '';
-                }
             }
 
             for (const e in this.errors) {
